rename to edge-containers-cli
