+++ .github/CONTRIBUTING.md
@@ -1,14 +1,14 @@
 # Contribute to the project
 
 Contributions and issues are most welcome! All issues and pull requests are
-handled through [GitHub](https://github.com/epics-containers/ec-cli/issues). Also, please check for any existing issues before
+handled through [GitHub](https://github.com/epics-containers/edge-containers-cli/issues). Also, please check for any existing issues before
 filing a new one. If you have a great idea but it involves big changes, please
 file a ticket before making a pull request! We want to make sure you don't spend
 your time coding something that might not fit the scope of the project.
 
 ## Issue or Discussion?
 
-Github also offers [discussions](https://github.com/epics-containers/ec-cli/discussions) as a place to ask questions and share ideas. If
+Github also offers [discussions](https://github.com/epics-containers/edge-containers-cli/discussions) as a place to ask questions and share ideas. If
 your issue is open ended and it is not obvious when it can be "closed", please
 raise it as a discussion instead.
