+++ README.md
@@ -1,9 +1,9 @@
-[![CI](https://github.com/epics-containers/ec-cli/actions/workflows/ci.yml/badge.svg)](https://github.com/epics-containers/ec-cli/actions/workflows/ci.yml)
-[![Coverage](https://codecov.io/gh/epics-containers/ec-cli/branch/main/graph/badge.svg)](https://codecov.io/gh/epics-containers/ec-cli)
-[![PyPI](https://img.shields.io/pypi/v/ec-cli.svg)](https://pypi.org/project/ec-cli)
+[![CI](https://github.com/epics-containers/edge-containers-cli/actions/workflows/ci.yml/badge.svg)](https://github.com/epics-containers/edge-containers-cli/actions/workflows/ci.yml)
+[![Coverage](https://codecov.io/gh/epics-containers/edge-containers-cli/branch/main/graph/badge.svg)](https://codecov.io/gh/epics-containers/edge-containers-cli)
+[![PyPI](https://img.shields.io/pypi/v/edge-containers-cli.svg)](https://pypi.org/project/edge-containers-cli)
 [![License](https://img.shields.io/badge/License-Apache%202.0-blue.svg)](https://opensource.org/licenses/Apache-2.0)
 
-# ec_cli
+# edge_containers_cli
 
 Edge containers command line interface: a tool for deploying and managing
 IOCs and services in a beamline or accelerator cluster.
@@ -12,15 +12,15 @@ For detailed documentation on the framework that this supports see:
 https://epics-containers.github.io/
 
 
-Source          | <https://github.com/epics-containers/ec-cli>
+Source          | <https://github.com/epics-containers/edge-containers-cli>
 :---:           | :---:
-PyPI            | `pip install ec-cli`
-Releases        | <https://github.com/epics-containers/ec-cli/releases>
+PyPI            | `pip install edge-containers-cli`
+Releases        | <https://github.com/epics-containers/edge-containers-cli/releases>
 
 
 To discover the CLI commands, create a python virtual environment,
 and perform the following::
 
-    pip install ec-cli
+    pip install edge-containers-cli
     ec --version
