+++ catalog-info.yaml
@@ -1,8 +1,8 @@
 apiVersion: backstage.io/v1alpha1
 kind: Component
 metadata:
-  name: ec-cli
-  title: ec-cli
+  name: edge-containers-cli
+  title: edge-containers-cli
   description: CLI for deploying and managing epics containers IOCs and services
 spec:
   type: documentation
