+++ docs/developer/tutorials/dev-install.rst
@@ -10,7 +10,7 @@ Clone the repository
 First clone the repository locally using `Git
 <https://git-scm.com/downloads>`_::
 
-    $ git clone git://github.com/epics-containers/ec-cli.git
+    $ git clone git://github.com/epics-containers/edge-containers-cli.git
 
 Install dependencies
 --------------------
@@ -25,7 +25,7 @@ requires python 3.8 or later) or to run in a container under `VSCode
 
         .. code::
 
-            $ cd ec-cli
+            $ cd edge-containers-cli
             $ python3 -m venv venv
             $ source venv/bin/activate
             $ pip install -e '.[dev]'
@@ -34,7 +34,7 @@ requires python 3.8 or later) or to run in a container under `VSCode
 
         .. code::
 
-            $ code ec-cli
+            $ code edge-containers-cli
             # Click on 'Reopen in Container' when prompted
             # Open a new terminal
