+++ docs/user/how-to/run-container.rst
@@ -1,15 +1,15 @@
 Run in a container
 ==================
 
-Pre-built containers with ec-cli and its dependencies already
+Pre-built containers with edge-containers-cli and its dependencies already
 installed are available on `Github Container Registry
-<https://ghcr.io/epics-containers/ec-cli>`_.
+<https://ghcr.io/epics-containers/edge-containers-cli>`_.
 
 Starting the container
 ----------------------
 
 To pull the container from github container registry and run::
 
-    $ docker run ghcr.io/epics-containers/ec-cli:main --version
+    $ docker run ghcr.io/epics-containers/edge-containers-cli:main --version
 
 To get a released version, use a numbered release instead of ``main``.
