+++ docs/user/tutorials/installation.rst
@@ -25,14 +25,14 @@ Installing the library
 
 You can now use ``pip`` to install the library and its dependencies::
 
-    $ python3 -m pip install ec-cli
+    $ python3 -m pip install edge-containers-cli
 
 If you require a feature that is not currently released you can also install
 from github::
 
-    $ python3 -m pip install git+https://github.com/epics-containers/ec-cli.git
+    $ python3 -m pip install git+https://github.com/epics-containers/edge-containers-cli.git
 
 The library should now be installed and the commandline interface on your path.
 You can check the version that has been installed by typing::
 
-    $ ec-cli --version
+    $ edge-containers-cli --version
