+++ pyproject.toml
@@ -3,7 +3,7 @@ requires = ["setuptools>=64", "setuptools_scm[toml]>=6.2"]
 build-backend = "setuptools.build_meta"
 
 [project]
-name = "ec-cli"
+name = "edge-containers-cli"
 classifiers = [
     "Development Status :: 3 - Alpha",
     "License :: OSI Approved :: Apache Software License",
@@ -44,10 +44,10 @@ dev = [
 ]
 
 [project.scripts]
-ec = "ec_cli.__main__:cli"
+ec = "edge_containers_cli.__main__:cli"
 
 [project.urls]
-GitHub = "https://github.com/epics-containers/ec-cli"
+GitHub = "https://github.com/epics-containers/edge-containers-cli"
 
 [[project.authors]] # Further authors may be added by duplicating this section
 email = "giles.knap@gmail.com"
@@ -55,7 +55,7 @@ name = "Giles Knap"
 
 
 [tool.setuptools_scm]
-write_to = "src/ec_cli/_version.py"
+write_to = "src/edge_containers_cli/_version.py"
 
 [tool.pyright]
 reportMissingImports = false # Ignore missing stubs in imported modules
@@ -75,7 +75,7 @@ filterwarnings = [
 testpaths = "src tests"
 
 [tool.coverage.run]
-data_file = "/tmp/ec_cli.coverage"
+data_file = "/tmp/edge_containers_cli.coverage"
 
 [tool.coverage.paths]
 # Tests are run from installed location, map back to the src directory
@@ -97,7 +97,7 @@ allowlist_externals =
     pre-commit
     pyright
 commands =
-    pytest: pytest --cov=ec_cli --cov-report term --cov-report xml:cov.xml {posargs}
+    pytest: pytest --cov=edge_containers_cli --cov-report term --cov-report xml:cov.xml {posargs}
     pyright: pyright src tests {posargs}
     pre-commit: pre-commit run --all-files {posargs}
 """
@@ -105,7 +105,7 @@ commands =
 [tool.ruff]
 src = ["src", "tests"]
 line-length = 88
-ignore = [
+lint.ignore = [
     "C408", # Unnecessary collection call - e.g. list(...) instead of [...]
     "E501", # Line too long, should be fixed by black.
     "B008", # Do not perform function calls in argument defaults.
